Implemented Terrain to do serial/deserial
This is a second approach to deserialization I have. Basically, it's the
more polymorphic/OOP approach. Let Terrain, the superclass of Floor and
Wall, do serializing/deserializing. Floor and Wall are indistinguishable
from each other except for their methods and sprites- neither of which
is considered in serialization. The fields that are serialized by all
Terrain subclasses should, logically, be serialized by Terrain!

Important note- Subclasses of Terrain definitely need to implement
deserialize (because we otherwise can't tell what type of terrain we
deserialized). Subclasses that are only distinguished by sprites and
methods need not implement serialize, since all of that work is done by
Terrain, but they DO need to implement it if they have specific member
variables that Terrain doesn't have. They still should make a super call
to Terrain's serialize to get all of Terrain's member variables out of
the way.