Important feature update (see desc)
- Click drag screen movement
- Clamped boundaries of screen so you can't move the level fully out of
screen
- *Added new UI object, a prompt, that is designed to run on a separate
thread and acquire user input from buttons.
- Started implementing serialization/deserialization. Floor is fully
implemented.
- Added a collections sorter to Terrain, that will sort lists of
applicable Terrain subclass instances by height, tallest first. (This is
important so that taller objects never completely obscure shorter ones-
same height ones, however, do)

- Layed groundwork for a robust "sort" in the level. The goal is to keep
everything in the level in one LevelObject list, in the order it needs
to be for rendering. Its less important for serializing, but when its
unpackaged in the actual game there will be no need to try and sort
stuff, it comes in order from the level format.
- Did I mention I can load levels?
- *Changed the layer system slightly- any OOP approach to that would
require a rewrite of my entire UI library so Im putting it off and
writing crap code

* So, I've had a lot of struggle with OOP this project, which is one
reason I'm taking it on. The idea of a "layer" is really throwing me
off. The older layers, for UI, render with a Graphics object, but the
ones Im using for literally everything else uses a Screen. So I'm
starting to think, should a screen be alayer? Im not sure, because it is
just a rendering device- which is layered, Ill grant you. So there is a
conflictual relationship between objects that render with the screen and
ones that don't. I can't keep them in the same list, unless I made a
super interface to combine them (easy enough) but its not very nifty.

Really I need to revisit the idea of what makes a layer, a layer.
Starting to think it needs to be dialed back to something that only
intercepts/handles events instead of also managing updates/renders as it
is starting to feel like Im losing the separation of concerns principle.
I've probably already lost it along the way but I'm trying damnit.

Anyway the project is coming along relatively smoothly, aside from a
couple bumps when I tackled trying to figure out the best ways to let
classes/super classes handle (de)serialization, its mostly all good and
I feel like I've accomplished learned a lot. Thanks invisible reader,
and see you next commit.