Lots of changes
-Added ability to reload and change assets (through the UI and a
shortcut key)
-Added save shortcut key (ctrl-s)
-Changed serialization/deserialization to reflect more of the level meta
stuff (width, height, level type, and if type is default)
- Gee I reallyfeel like I did a lot more :camel:
- Messed around with assets loading, a lot.  There are like three
methods, and my naming convention isnt the greatest, so I guarantee
there is a bit of inefficiency. Also a nasty bug that I tried
circumventing, ifyou try loading the assets for a level type that
doesn't have all of the folders. Made me give everything a bunch of
throw IOExceptions, and handled them at the start calling points which
was a pain. Replaced the bad file with the default Fields.
- Updated file loading to not use class.getresource, reources are now
independently situated outside the jar for ease of editing. The big
problem with this and why I changed it is because it was not syncing up
whenever I made edits and tried reloading assets, so that all works now.