Lots more changes
- All level objects should now be getting access to the level that owns
them, through init. This is for helping determine the boundaries of
placement, so we can't move shit out side of the level.
I will need to be careful when levels are shortened- either delete those
things or move them in.
- Started working on click drag movement, went very smoothly. No
problems here, just needs to be implemented with the rest of the more
abstract level objects when we actually get around to making those.
- There is now a fairly comprehensive saving/loading system. I have
chronic being-bad disease, so the main class, editor, is looking pretty
spaghetti. I really need to work on separation of concerns, in this
particular part of my project. There are too manyexceptions to how I
save and load, i.e when I close the program I save but I dont save in
the same way that I do when I hit ctrl-s. It works, but it's not clean.
- Minor change to my approach to object rendering- Floors now determine
the "width" to render their trimmings (sides and top) based on the
actual width (or top in the case of the top) of each sprite. This way it
still wont tile in the inappropriate directions, but it will be the
appropriate height/width that was drawn.
- Level now has a reference to the editor, for access to those sweet
sweet methods/members.
- Levels, upon deserializing, will now use the file name in place of
what is stored in the database, so you can change the name of the level
while not in the editor. There is currently no way to change the name IN
the editor.
- Added actual event capturing to the level... Its a pain in the ass
working with the event thread.
- Whenever a level is loaded after the initial startup, it first
refreshes the level directory to account for updates