Added more editing features
- Level objects now have an optional property panel, on the bottom. The
property panel comes up when you click on an object and it contains text
fields to edit width and height. They aren't labelled.
- Level width and height is also now editable from the left hand editor
panel, which is always present.
- Also on the editor panel is a new scroll list. The pane lets you
choose which level object you want to place, and then you place it by
middle clicking. I went through a lot of pain trying to do this because
I thought generics would help me but I settled on enums and a factory
class.
- Implemented Walls alongside floors. Both are fully implemented, all we
need is whatever other level things we want.

The level editor is mostly done. The framework for
editing/loading/saving levels is entirely in place. All that it needs is
for the rest of the level objects to be implemented. Labels and more UI
options would be nice, too. Its not very user friendly but it is Henry
friendly.