Click dragging objects actually drags the object
In practice, the translated movements of the mouse after clicking an
object are applied to the object. This replaces the objects x,y just
being set to the mouse's.

What's going on under the hood is I'm setting distance variables
everytime the object is clicked. When the object is clicked, distance
from the upper left corner (where the object is located and draws from)
to the mouse is set. The X and Y distances are then applied to the X and
Y of wherever the mouse moves to, "dragging" the object relatively along
with it, such that distance from the object's actual x and y to the
mouse is preserved.

Also minor comments.

Side note, because the Y axis is inverted (because reasons :+1: ), the y
distance modifier is added instead of subtracted. Since we're guaranteed
that the mouse is always technically to the right of the object X, and
below the object Y (where normally it would be above) we don't have to
do additional calculating to figure out which way to apply the distance
modifiers, we just know that we need to add y and subtract x.

Like wise in initially calculating the difference, Mouse X subtracts
Object X, but Object Y subtracts Mouse Y. I really need to create a
"difference" math utility function! It would make the setting of the
modifiers more readable!